feature hygiene: a flag answers 'what platform', never 'which guarantees' — 4 footguns closed + a vacuous gate revived
The principle enforced (docs/FEATURE-HYGIENE.md, 46 features / 24 crates
censused): platform/perf gates (wasm32, SP1 guest, native prover, …) KEPT (22);
semantic-optionality is a footgun — you must not be able to compile a dregg
that silently lacks a guarantee.

CLOSED (made always-on, core semantics): wire/stark-verifier + wire/bridge (the
STARK + bridge verifiers, fail-closed stub deleted — optionality was illusory,
dregg-dsl-runtime already pulled plonky3); bridge/turn (turn-proof verifier);
persist/audit-bridge (an auditable dregg can no longer be compiled unauditable).

BEST FIND — a vacuous gate revived: preflight::check_plonky3_backend was
wrapped in #[cfg(feature = "plonky3")] but preflight declares NO such feature,
so the cfg was ALWAYS FALSE and the preflight backend check PASSED VACUOUSLY.
Now unconditional and real. (+ bridge present.rs bench->test-utils cfg fix.)

DELETED dead: circuit/mock, circuit/dev, wire/dev (gated nothing) + the
dangling enables across wasm/sdk/apps.

FLAGGED for ember (one decision each, named not logged): the 7 Lean-gate
features bottom out in linking libdregg_lean.a (a real native-link constraint
for wasm/zkvm) — recommended closure = invert polarity (Lean unconditional on
native, one platform-no-lean-link gate); chain/mock as DEFAULT; dfa verifier
compiled into no build. (persist/commit_log test helper: a signed-wells i64
straggler, fixed.)

cargo check green on persist/bridge/wire/sdk/preflight/circuit + wasm32 target;
nothing reverted.

+++ persist/src/commit_log.rs
@@ -793,8 +793,11 @@ mod tests {
         }
     }
 
+    // signed-wells (ac01f9b7b): cell balances are i64; this test helper keeps a
+    // u64 convenience param (callers pass small non-negative amounts) and
+    // converts at the boundary.
     fn cell(seed: u8, balance: u64) -> Cell {
-        Cell::with_balance([seed; 32], [seed.wrapping_add(7); 32], balance)
+        Cell::with_balance([seed; 32], [seed.wrapping_add(7); 32], balance as i64)
     }
 
     #[test]
